Added Creep armor

+++ Assets/2. Scripts/Creeps/Health.cs
@@ -1,17 +1,17 @@
 ﻿namespace Core {
     public class Health {
-        public int current { get; private set; }
-        public int max { get; private set; }
+        public float current { get; private set; }
+        public float max { get; private set; }
 
-        public Health(int amnt) {
+        public Health(float amnt) {
             max = current = amnt;
         }
 
-        public void SetCurrent(int value) {
+        public void SetCurrent(float value) {
             current = value;
         }
 
-        public void DealDamage(int amnt) {
+        public void DealDamage(float amnt) {
             current -= amnt;
 
             if (current  < 0) {
@@ -19,7 +19,7 @@
             }
         }
 
-        public void AddHealth(int amnt) {
+        public void AddHealth(float amnt) {
             current += amnt;
             if (current > max) {
                 current = max;
@@ -28,7 +28,7 @@
         public float Ratio() {
             return (float)current / max;
         }
-        public void Reset(int max = -1) {
+        public void Reset(float max = -1) {
             if (max <= 0) {
                 max = this.max;
             }
